support MSD dataset.json, dump training dataset in output_dir
