+++ src/segmantic/seg/monai_unet.py
@@ -45,6 +45,7 @@ from monai.utils import set_determinism
 from pytorch_lightning.callbacks import ModelCheckpoint
 from pytorch_lightning.loggers import TensorBoardLogger
 
+from ..util.encoders import PathEncoder
 from ..prepro.labels import load_tissue_list
 from .dataset import PairedDataSet
 from .evaluation import confusion_matrix
@@ -332,6 +333,17 @@ def train(
     net.intensity_augmentation = augment_intensity
     net.spatial_augmentation = augment_spatial
     net.cache_rate = cache_rate
+
+    # store dataset used for training
+    (output_dir / "Dataset.json").write_text(
+        json.dumps(
+            {
+                "trainig": net.dataset.training_files(),
+                "validation": net.dataset.validation_files(),
+            },
+            cls=PathEncoder,
+        )
+    )
 
     # set up loggers and checkpoints
     tb_logger = TensorBoardLogger(save_dir=f"{log_dir}")
